pisces-api/pyproject.toml
[project]
name = "pisces-api"
version = "0.1.0"
description = "Chinese → Vietnamese novel translation pipeline (multi-agent, LangGraph)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "celery[sqlalchemy]>=5.6.3",
    "ebooklib>=0.20",
    "fastapi[standard]>=0.136.0",
    "langchain-core>=1.2.31",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0.49",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.pyright]
venvPath = "."
venv = ".venv"

